create MyComment commponet
